requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
mediapipe
numpy
pillow
pydantic-settings
gradio